app/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.0
Django==5.0.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
